backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ruben <ruben23fg@gmail.com>"]
readme = ""

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
sqlalchemy = "^2.0.43"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"
passlib = "^1.7.4"
jwt = "^1.4.0"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
